+++ srcs/zoom.c
@@ -6,13 +6,13 @@
 /*   By: zytrams <zytrams@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/18 19:30:25 by zytrams           #+#    #+#             */
-/*   Updated: 2019/06/06 16:57:53 by zytrams          ###   ########.fr       */
+/*   Updated: 2019/11/27 05:01:32 by zytrams          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <fdf.h>
 
-void				zoom_in(double *scalex, double *scaley, double *scalez)
+void				zoom_in(float *scalex, float *scaley, float *scalez)
 {
 	if (*scalex < 4 && *scaley < 4 && *scalez < 4)
 	{
@@ -22,7 +22,7 @@ void				zoom_in(double *scalex, double *scaley, double *scalez)
 	}
 }
 
-void				zoom_out(double *scalex, double *scaley, double *scalez)
+void				zoom_out(float *scalex, float *scaley, float *scalez)
 {
 	*scalex /= 2;
 	*scaley /= 2;
